made the TopSellers information dynamic
