+++ src/pages/Home.jsx
@@ -1,36 +1,15 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import BrowseByCategory from "../components/home/BrowseByCategory";
 import HotCollections from "../components/home/HotCollections";
 import Landing from "../components/home/Landing";
 import LandingIntro from "../components/home/LandingIntro";
 import NewItems from "../components/home/NewItems";
 import TopSellers from "../components/home/TopSellers";
-import axios from "axios";
 
 const Home = () => {
   useEffect(() => {
     window.scrollTo(0, 0);
   }, []);
-
-  const [collections, setCollections] = useState([]);
-    const [loading, setLoading] = useState(true);
-
-  async function fetchCollections() {
-    try {
-      const { data } = await axios.get(
-        "https://us-central1-nft-cloud-functions.cloudfunctions.net/hotCollections"
-      );
-      setCollections(data);
-      setLoading(false);
-    } catch (error) {
-      console.error("Failed to fetch collections:", error);
-      setLoading(false);
-    }
-  }
-
-  useEffect(() => {
-      fetchCollections();
-    }, []);
 
   return (
     <div id="wrapper">
